chyba skonczone jeszcze tylko komentarze

+++ meteo.sh
@@ -28,28 +28,50 @@ else
     homeCity=$1
 fi
 
+if ! [[ -e "$HOME/.meteorc" ]] # sprawdzenie pliku konfiguracyjnego lub stworzenie go z podstawowymi wartosciami
+then
+    echo 'No cache file'
+    echo '{"cashpath":"'"$HOME"'/.cache/meteo/"}' > "$HOME/.meteorc"
+    mkdir -p "$HOME/.cache/meteo/"
+fi
 
-cityCashPath='./cityCash.json'
+cashPath=$(cat "$HOME/.meteorc" | jq -r '.cashpath')
 
 if [[ $debug -eq 1 ]]
 then
 echo 'pobieranie danych z IMGW'
 fi
-meteoAPI=$(curl -s https://danepubliczne.imgw.pl/api/data/synop) # pobieranie danych od IMGW-PIB
-if test -e $cityCashPath;
+
+if [[ -e $cashPath'meteoCash.json' && -e $cashPath'meteoCash.txt' ]] # pobieranie danych z IMGW tylko gdy dane sa starsze niz godzina lub gdy ich nie ma
 then
-    cityCash=$(cat $cityCashPath)
+    meteoAPI=$(cat $cashPath'meteoCash.json')
+    if [[ "$(echo $meteoAPI | jq -r '.[0].data_pomiaru')$(echo $meteoAPI | jq -r '.[0].godzina_pomiaru')" != "$(cat $cashPath'meteoCash.txt')" ]]
+    then
+        meteoAPI=$(curl -s https://danepubliczne.imgw.pl/api/data/synop) # pobieranie danych od IMGW-PIB
+        echo $meteoAPI > $cashPath'meteoCash.json'
+        echo "$(echo $meteoAPI | jq -r '.[0].data_pomiaru')$(echo $meteoAPI | jq -r '.[0].godzina_pomiaru')" > $cashPath'meteoCash.txt'
+    fi
+else
+    meteoAPI=$(curl -s https://danepubliczne.imgw.pl/api/data/synop) # pobieranie danych od IMGW-PIB
+    echo $meteoAPI > $cashPath'meteoCash.json'
+    echo "$(echo $meteoAPI | jq -r '.[0].data_pomiaru')$(echo $meteoAPI | jq -r '.[0].godzina_pomiaru')" > $cashPath'meteoCash.txt'
+fi
+
+
+
+if test -e $cashPath'cityCash.json';
+then
+    cityCash=$(cat $cashPath'cityCash.json')
 else
     echo 'pierwsze uruchomienie potrwa ponad minute z powodu ograniczen nalozonych przez https://nominatim.org/'
     echo 'prosze o cierpliwosc :)'
-    echo {} > $cityCashPath
+    echo {} > $cashPath'cityCash.json'
 fi
-
 
 # pobieranie danych z api nominatim
 for ((i=0; i<=$(echo $meteoAPI | jq '. | length'); i++));
 do
-
+    echo -n '*'
     if [[ $i -eq $(echo $meteoAPI | jq '. | length') ]]
     then
         cityName=$1
@@ -71,30 +93,37 @@ do
         then
         echo 'pobieranie i zapisanie danych z nominatim dla '$cityName
         fi
+
         cityGeometry=$(curl -s -L -H "User-Agent: Mozilla/5.0" "https://nominatim.openstreetmap.org/search?country=Poland&city='$cityName'&limit=1&format=geojson" | jq '.features[0].geometry')
 
-        echo $cityGeometry > temp.json
-        echo $(jq --argfile cityGeometry temp.json '.'$cityName' += $cityGeometry' $cityCashPath) > $cityCashPath
-        rm temp.json
-
+        if [[ -z "$cityGeometry" || "$cityGeometry" == "null" ]]
+        then
+            echo "Error while fetching data from Nominatim for $cityName"
+            exit 1
+        else
+            echo $cityGeometry > temp.json
+            echo $(jq --argfile cityGeometry temp.json '.'$cityName' += $cityGeometry' $cashPath'cityCash.json') > $cashPath'cityCash.json'
+            rm temp.json
+        fi
         sleep 1 # nominatim przyjmuje max 1 zapytanie na 1s
     fi
 done
 
-
+echo ' '
 homeCity=$(echo $homeCity | iconv -f utf8 -t ascii//TRANSLIT) #usuwanie polskich znakow z nazwy miasta
 homeCity=$(echo $homeCity | tr ' ' '_') #usuwanie spacji z nazwy miasta
 
-inputx=$(cat cityCash.json | jq '.'$homeCity'.coordinates[0]')
-inputy=$(cat cityCash.json | jq '.'$homeCity'.coordinates[1]')
+inputx=$(cat $cashPath'cityCash.json' | jq '.'$homeCity'.coordinates[0]')
+inputy=$(cat $cashPath'cityCash.json' | jq '.'$homeCity'.coordinates[1]')
 
 shortCity=$(echo $meteoAPI | jq -r '.[0].stacja')
 
 for ((i=0; i<$(echo $meteoAPI | jq '. | length'); i++)); # ustalenie najblizszego miasta
 do
+    echo -n '*'
     shortCityPars=$(echo $shortCity | iconv -f utf8 -t ascii//TRANSLIT | tr ' ' '_') #usuwanie polskich znakow i spacji z nazwy miasta
-    shortx=$(cat cityCash.json | jq '.'$shortCityPars'.coordinates[0]')
-    shorty=$(cat cityCash.json | jq '.'$shortCityPars'.coordinates[1]')
+    shortx=$(cat $cashPath'cityCash.json' | jq '.'$shortCityPars'.coordinates[0]')
+    shorty=$(cat $cashPath'cityCash.json' | jq '.'$shortCityPars'.coordinates[1]')
 
     # obliczanie odleglosci od poprzedniego najblizszego miasta
     deltax=$(echo $inputx-$shortx | bc)
@@ -104,8 +133,8 @@ do
     cityName=$(echo $meteoAPI | jq -r '.['$i'].stacja')
     cityNamePars=$(echo $cityName | iconv -f utf8 -t ascii//TRANSLIT | tr ' ' '_') #usuwanie polskich znakow i spacji z nazwy miasta
     
-    cityx=$(cat cityCash.json | jq '.'$cityNamePars'.coordinates[0]')
-    cityy=$(cat cityCash.json | jq '.'$cityNamePars'.coordinates[1]')
+    cityx=$(cat $cashPath'cityCash.json' | jq '.'$cityNamePars'.coordinates[0]')
+    cityy=$(cat $cashPath'cityCash.json' | jq '.'$cityNamePars'.coordinates[1]')
 
     # obliczanie odleglosci od miasta
     deltax=$(echo $inputx-$cityx | bc)
@@ -128,6 +157,7 @@ done
 #shortCityIdx=$(echo $meteoAPI | jq 'map(.stacja == '\"$shortCity\"') | index(true)')
 shortCityIdx=$(echo $meteoAPI | jq --arg shortCity "$shortCity" 'map(.stacja == $shortCity) | index(true)')
 
+echo ' '
 
 echo $shortCity '['$(echo $meteoAPI | jq -r '.['$shortCityIdx'].id_stacji')'] /' $(echo $meteoAPI | jq -r '.['$shortCityIdx'].data_pomiaru') $(echo $meteoAPI | jq -r '.['$shortCityIdx'].godzina_pomiaru')":00"
 echo "temperatura:" $(echo $meteoAPI | jq -r '.['$shortCityIdx'].temperatura') "°C"
